+++ ferry/lib/cache.dart
@@ -1,4 +1,5 @@
 export 'package:ferry_store/ferry_store.dart';
+export 'package:normalize/policies.dart';
 
 export './src/cache.dart';
 export './src/fetch_policy.dart';