+++ ferry/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.3] - 2020-09-09
+
+### Changed
+- Bump `normalize` version
+- Export policies in ferry and cache
+
 ## [0.7.2] - 2020-09-09
 
 ### Added